Create a new facility detail sidebar
In order to allow maintenance work to be released while the Extended
Profile feature is in development, the new sidebar is wrapped in a
feature flag. When the flag is disabled, the old sidebar is displayed.

This initial implementation of the ExtendedProfile sidebar has only a
limited subset of fields and behaviors implemented. Specifically,
the claim status header; the OAR ID; the facility name, address, and
location; the static map; the contributors; and the action buttons have
been implemented.

Extended fields, claim fields, contributor/nonstandard fields, the
closed facility status and activity information, and the 'additional
information' popups have not yet been implemented.
